requirements-ui.txt
tkinterdnd2
fonttools
Pillow
numpy
pdf2image
tensorflow
